pyproject.toml
[project]
name = "hermes3d-os"
version = "0.1.0"
description = "Standalone Hermes-powered 3D print factory controller."
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.6",
  "httpx==0.28.1",
  "pydantic==2.10.5",
  "python-multipart==0.0.20",
  "PyYAML==6.0.2",
  "uvicorn[standard]==0.34.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

